Can specify that all SAN names be encoded as dNSName.
Micros expects all SAN names to be dNSNames. r509 will conveniently
detect if a SAN name is an IP address and encode it as an iPAddress, but
we want to be able to tell it to not do that. The general case remains
the same, but now if you pass extensions[dNSNames][] instead of
extensions[subjectAlternativeName][], all the SAN names will be encoded
as dNSNames.